pr corrections
